refactor: path dos exceptions de aluno
