+++ src/main/java/br/edu/ufape/sgi/servicos/interfaces/AlunoService.java
@@ -1,6 +1,6 @@
 package br.edu.ufape.sgi.servicos.interfaces;
 
-import br.edu.ufape.sgi.exceptions.aluno.AlunoNotFoundException;
+import br.edu.ufape.sgi.exceptions.notFoundExceptions.AlunoNotFoundException;
 import br.edu.ufape.sgi.models.Aluno;
 
 import java.util.List;
